@@ -12,43 +12,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
+    "@angular/animations": "^11.2.8",
+    "@angular/common": "^11.2.8",
+    "@angular/compiler": "^11.2.8",
+    "@angular/core": "^11.2.8",
+    "@angular/forms": "^11.2.8",
+    "@angular/platform-browser": "^11.2.8",
+    "@angular/platform-browser-dynamic": "^11.2.8",
+    "@angular/router": "^11.2.8",
     "bootstrap": "^4.1.1",
     "core-js": "^2.4.1",
     "jquery": "^3.3.1",
     "ng-bootstrap": "^1.6.3",
     "popper.js": "^1.14.3",
-    "rxjs": "^5.5.10",
-    "zone.js": "^0.8.19"
+    "rxjs": "^6.6.7",
+    "tslib": "^2.0.0",
+    "zone.js": "~0.10.2",
+    "rxjs-compat": "^6.0.0-rc.0"
   },
   "devDependencies": {
-    "@angular/cli": "~1.7.4",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "@types/jasmine": "~2.8.3",
+    "@angular-devkit/build-angular": "~0.1102.7",
+    "@angular/cli": "^11.2.7",
+    "@angular/compiler-cli": "^11.2.8",
+    "@angular/language-service": "^11.2.8",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^6.0.0",
     "concurrently": "^3.5.1",
     "express": "^4.16.3",
-    "jasmine-core": "~2.8.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.2",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.3.2",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
     "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "tslint": "~6.1.0",
+    "typescript": "~4.0.7"
   }
 }
